Turn sub h1s into h2s + beyond
In older HTML, pages shouldn't have more than h1, and they should
follow the single hierarchy defined by descending headings. In newer
HTML, you're allowed to have multiple h1s if they're in specific
sectioning content like section and aside.

I originally used the multiple h1s on the page approach because I liked
the idea of a poem transfomation that takes the text and transform it
without context into something like

```
<aside>
  <h1>Title</h1>

  <p>
    The poem<br />
    is so good
  </p>
</aside>
```

If a page expected this to be a nested child, I would need to pass in
the expected depth in order to get the correct heading level.

The reason I switched this to the single-hierarchy approach is because
many clients don't know how to interpret the newer multiple-h1 style. I
wanted something more accessible to machines. My browser adjusted
heading size based on nested sections, but I imagine others do not.
Futhermore, I'm not even sure I know all the rules as to what creates a
new section.

I might change my mind on this later, but the codebase is small enough
that I don't mind looking over the headings for a single hierarchy.